src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/getelementbyid-shadow-expected.txt

CONSOLE ERROR: line 23: Uncaught TypeError: A.createShadowRoot is not a function
Tests to ensure ShadowRoot.getElementById works after complex DOM manipulation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
